+++ src/TextWriter.h
@@ -35,7 +35,7 @@ public:
 	TextWriter(const GlobalData *_gdata);
 	~TextWriter();
 
-	virtual void write(uint numParts, BufferList const& buffers, uint node_offset, float t, const bool testpoints);
+	virtual void write(uint numParts, BufferList const& buffers, uint node_offset, double t, const bool testpoints);
 };
 
 #endif
