+++ src/UDPWriter.h
@@ -45,7 +45,7 @@ public:
 	UDPWriter(const GlobalData *_gdata);
 	~UDPWriter();
 
-	virtual void write(uint numParts, BufferList const& buffers, uint node_offset, float t, const bool testpoints);
+	virtual void write(uint numParts, BufferList const& buffers, uint node_offset, double t, const bool testpoints);
 
 protected:
     double3     mWorldOrigin,
