+++ src/VTKLegacyWriter.h
@@ -36,7 +36,7 @@ public:
 	VTKLegacyWriter(const GlobalData *_gdata);
 	~VTKLegacyWriter();
 
-	virtual void write(uint numParts, BufferList const& buffers, uint node_offset, float t, const bool testpoints);
+	virtual void write(uint numParts, BufferList const& buffers, uint node_offset, double t, const bool testpoints);
 };
 
 #endif	/* _VTKLEGACYWRITER_H */
